% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{apply_poppk}
\alias{apply_poppk}
\title{Apply a population PK model to a data frame}
\usage{
apply_poppk(patient_df, mod = c("marsh", "schnider", "eleveld"), ...)
}
\arguments{
\item{patient_df}{Dataframe with patient covariate values. Must have
names used by model "mod"}

\item{mod}{Population PK model to apply to rows of patient_df}

\item{...}{Arguments passed on to population PK model.}
}
\value{
The function applies either the Marsh (1991), Schnider et al. (1998), or Eleveld et al (2018) PK models for
propofol to a dataframe with patient covariate values and returns PK (or PK-PD, in the case of the Eleveld model)
parameter values. The function returns the original dataframe merged with parameter values for each individual (in rows).
}
\description{
Function to apply saved population PK or PK-PD models to a
data frame of patient values.
}
