% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpd_assist.R
\name{assign_pars}
\alias{assign_pars}
\title{Set default PK parameter values
Set default PK parameter values for a pkmod object.}
\usage{
assign_pars(pkmod, pars)
}
\arguments{
\item{pkmod}{pkmod object}

\item{pars}{PK parameters to assign as default values of pkmod}
}
\value{
pkmod object
}
\description{
Set default PK parameter values
Set default PK parameter values for a pkmod object.
}
