% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{cl_targets}
\alias{cl_targets}
\title{Closed-loop targets}
\usage{
cl_targets(time, target)
}
\arguments{
\item{time}{Times at which target values are set}

\item{target}{Response target values}
}
\value{
Returns a data frame with columns "time" and "target" as are required for use in the "bayes_control" function.
}
\description{
Format data frame of closed-loop targets.
}
