% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{cl_updates}
\alias{cl_updates}
\title{Closed-loop updates}
\usage{
cl_updates(time, full_data = TRUE, plot_progress = FALSE)
}
\arguments{
\item{time}{Times at which PK or PK-PD parameters should be updated}

\item{full_data}{Vector of logical values indicating if all data should be used at update
or only data since last update. Once first "FALSE" value is observed, the prior variance-
covariance matrix is overwritten. Consequently, any "TRUE" updates will only use data since
the last "FALSE" update.}

\item{plot_progress}{Vector of logical values. Should values be plotted at each update?}
}
\description{
Set parameters for closed-loop updates.
}
