% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{combine_sim}
\alias{combine_sim}
\title{Combine simulation outputs}
\usage{
combine_sim(...)
}
\arguments{
\item{...}{Set of datasim objects created from `gen_data` function.}
}
\description{
Function to merge objects with class datasim from different infusion schedules
infusion schedules can be passed directly in or as a list.
}
