% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpd_assist.R
\name{create_inf}
\alias{create_inf}
\title{Create dosing schedule}
\usage{
create_inf(times, infrt = NULL, duration = NULL)
}
\arguments{
\item{times}{Vector of time to begin infusions. If duration is NULL, times
expected to include both infusion start and infusion end times.}

\item{infrt}{Vector of infusion rates. Must lave length equal to either
length(times) or 1. If length(infrt)=1, the same infusion rate will be
administered at each time. Either infrt or target must be specified, but not both.}

\item{duration}{Optional duration of infusions.}
}
\value{
Matrix of infusion rates, start and end times.
}
\description{
Returns a data frame describing a set of infusions to be administered. Output
can be used as argument "inf" in predict_pkmod.
}
\examples{
# specify start and end times, as well as infusion rates (including 0).
create_inf(times = c(0,0.5,4,4.5,10), infrt = c(100,0,80,0,0))
# specify start times, single infusion rate and single duration
create_inf(times = c(0,4), infrt = 100, duration = 0.5)
# multiple infusion rates, single duration
create_inf(times = c(0,4), infrt = c(100,80), duration = 0.5)
# multiple sequential infusion rates
create_inf(times = seq(0,1,1/6), infrt = 100, duration = 1/6)
# single infusion rate, multiple durations
create_inf(times = c(0,4), infrt = 100, duration = c(0.5,1))
# multiple infusion rates, multiple durations
create_inf(times = c(0,4), infrt = c(100,80), duration = c(0.5,1))
}
