% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{cubic_targetfn}
\alias{cubic_targetfn}
\title{cubic target function restricted to pass through BIS = 93 at t=0 and BIS = 50 at t=5}
\usage{
cubic_targetfn(pars, tms, bis0 = 93, tfinal = 10, bisfinal = 50)
}
\description{
cubic target function restricted to pass through BIS = 93 at t=0 and BIS = 50 at t=5
}
