% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{cut2}
\alias{cut2}
\title{cut2}
\usage{
cut2(
  x,
  breaks,
  labels = NULL,
  include.lowest = FALSE,
  right = TRUE,
  dig.lab = 3L,
  ordered_result = FALSE,
  ...
)
}
\description{
cut2
}
\examples{
tail_vec(1:8)
tail_vec(matrix(1:8,2,4))
Modified cut function

Modified cut function to create equal-length intervals. See documentation
for the base function 'cut' for description of function arguments. Not currently used.

}
