% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpd_assist.R
\name{elvdlpars}
\alias{elvdlpars}
\title{elvdlpars}
\usage{
elvdlpars(x, pd = TRUE)
}
\arguments{
\item{x}{Vector or data frame with Eleveld PK-PD model parameters}

\item{pd}{Logical. Should PD parameters be returned in addition to PK parameters.}
}
\description{
elvdlpars
}
\examples{
poppk_cov("Eleveld", pd = TRUE)
Get logged parameters updated in Eleveld model

Extract the logged parameter values to be updated within the Eleveld model
from a data frame of patient PK-PD values.

}
