% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_mods.R
\name{emax}
\alias{emax}
\title{Emax function. c50 is the concentration eliciting a 50% effect, gamma is the hill parameter
identifying the slope of the Emax curve at c50, E0 is the response value with no drug present,
Emx is the maximum effect size.}
\usage{
emax(ce, pars)
}
\arguments{
\item{ce}{Vector of effect-site concentrations.}

\item{pars}{Named vector of parameter values with names (c50,gamma,e0,emx).}
}
\description{
Emax function. c50 is the concentration eliciting a 50% effect, gamma is the hill parameter
identifying the slope of the Emax curve at c50, E0 is the response value with no drug present,
Emx is the maximum effect size.
}
