% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_mods.R
\name{gen_eleveld_pk_pars}
\alias{gen_eleveld_pk_pars}
\title{gen_eleveld_pk_pars}
\usage{
gen_eleveld_pk_pars(theta, eta, patient_vars, returnQ = FALSE)
}
\arguments{
\item{theta}{Vector of fixed effects}

\item{eta}{Vector of random effects}

\item{patient_vars}{Named list of observed patient characteristics}

\item{returnQ}{Logical. Should clearance be returned instead of rates}
}
\description{
gen_eleveld_pk_pars
}
\examples{
pk_pars <- eleveld_pk[eleveld_pk$ID == 403,c("V1","V2","V3","CL","Q2","Q3")]
pd_pars <- eleveld_pd[eleveld_pd$ID == 403,c("E50","KE0","EMAX","GAM","GAM1","RESD")]
pars <- c(k10 = pk_pars$CL / pk_pars$V1,
          k12 = pk_pars$Q2 / pk_pars$V1,
          k21 = pk_pars$Q2 / pk_pars$V2,
          k13 = pk_pars$Q3 / pk_pars$V1,
          k31 = pk_pars$Q3 / pk_pars$V3,
          v1 = pk_pars$V1,
          v2 = pk_pars$V2,
          v3 = pk_pars$V3,
          ke0 = pd_pars$KE0)
ivt <- list(begin = 0, end = 0.5, k_R = 1)
init <- c(0,0,0,0)
sol <- pk_solution_3cpt_metab_singleinf(pars, ivt, init)
tms <- seq(0,20,0.1)
plot(tms, sol(tms)[1,], type = "l", xlab = "Minutes", ylab = "Concentration")
lines(tms, sol(tms)[4,], col = 2)
legend("bottomright", c("Central concentration", "Effect-site concentration"), col = c(1,2), lty = c(1,1))
Eleveld model PK parameters

Function to generate PK parameters for Eleveld model.

}
