% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_mods.R
\name{inv_emax}
\alias{inv_emax}
\alias{emax_eleveld}
\title{inv_emax}
\usage{
inv_emax(pdresp, pars)

emax_eleveld(ce, pars)
}
\arguments{
\item{pdresp}{PD response values}

\item{pars}{Vector of parameter values in order (c50,gamma,gamma2,e0,emx).}

\item{ce}{Vector of effect-site concentrations.}
}
\description{
pars_emax <- c(c50 = 1.5, gamma = 4, e0 = 100, emx = 100)
ce_seq <- seq(0,4,0.1)
all.equal(inv_emax(emax(ce_seq, pars_emax), pars_emax), ce_seq)
Emax function for Eleveld (2018) model.
}
\details{
The parameter gamma takes one of two values depending on whether ce <= c50.
}
\examples{
pars_emax <- c(c50 = 1.5, gamma = 1.47, e0 = 100, emx = 100)
ce_seq <- seq(0,4,0.1)
plot(ce_seq, emax(ce_seq, pars_emax), type = "l", xlab = "Effect-site concentrtion (ug/mL)", ylab = "BIS")
Inverse Emax function
}
