% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpd_assist.R
\name{list_parnms}
\alias{list_parnms}
\title{Identify pkfn from parameter names}
\usage{
list_parnms()
}
\value{
Returns one of the following functions: `pkmod1cpt`, `pkmod2cpt`, `pkmod3cpt`,
or `pkmod3cptm` based on the parameter names entered.
}
\description{
Identify structural PK model function (i.e., `pkfn`) from parameter names.
Models available are 1-, 2-, and 3-compartment mammillary models, or 3-compartment with
an effect site, corresponding to functions `pkmod1cpt`, `pkmod2cpt`, `pkmod3cpt`, and
`pkmod3cptm`, respectively.
}
\examples{
list_parnms()
}
