% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpd_assist.R
\name{list_pkmods}
\alias{list_pkmods}
\title{Print population PK models available in `tci`}
\usage{
list_pkmods()
}
\value{
Prints function names, model types, and required covariates for each
model.
}
\description{
Print population PK models available in `tci` for propofol (Marsh,
Schnider, Eleveld) and remifentanil (Minto, Kim, Eleveld).
}
\examples{
list_pkmods()
}
