% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\docType{data}
\name{pal}
\alias{pal}
\title{Color palate for tci plotting functions}
\format{
An object of class \code{character} of length 7.
}
\usage{
pal
}
\description{
Color palate for tci plotting functions
}
\keyword{datasets}
