% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{Phi}
\alias{Phi}
\title{Objective function for a single infusion - numerically estimates the weighted integral between the time-BIS curve and the target BIS value.}
\usage{
Phi(
  kR,
  theta,
  E0,
  gamma,
  t0,
  alpha = 0.01,
  bist = 50,
  tfinal = 10,
  dt = 0.01,
  log = F
)
}
\arguments{
\item{kR}{Infusion schedule to be evaluated.}

\item{theta}{Vector of PK-PD parameters (theta_PK, theta_PD)}

\item{E0}{Effect at ce = 0. Used as starting point for sigmoid curve.}

\item{gamma}{Hill parameter (slope at c50) for PD model}

\item{alpha}{Weight associated with the integral above the target BIS value and below the time-BIS curve.}

\item{bist}{Target BIS, used to set the maximum effect value of the target sigmoid curve.}

\item{tfinal}{Final time to evaluate infusion schedule to.}

\item{dt}{Resolution of integral}
}
\description{
Objective function for a single infusion - numerically estimates the weighted integral between the time-BIS curve and the target BIS value.
}
