% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpd_assist.R
\name{poppk_cov}
\alias{poppk_cov}
\title{Get variance-covariance matrix for population PK-PD models}
\usage{
poppk_cov(poppk = c("Schnider", "Eleveld"), pd = TRUE)
}
\arguments{
\item{poppk}{Selection of Schnnider PK model or Eleveld PK-PD model}

\item{pd}{Logical. Should PD parameters be returned for Eleveld model}
}
\description{
NOTE: This needs to be modified. There are correlations between parameters
induced by patient covariates.
}
\details{
Function to extract covariance from population pk or pk-pd models
This gives the covariance for the random effects about the fixed effect estimates.
The variance estimates for the residual error terms give the variance of error terms
within the population about the logged
}
