% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pkmod}
\alias{print.pkmod}
\title{Print pkmod}
\usage{
\method{print}{pkmod}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object with class "pkmod" created by `pkmod()`}

\item{...}{Arguments passed to `update.pkmod()`}

\item{digits}{Number of significant digits to print}
}
\value{
Prints description of pkmod
}
\description{
Print method for pkmod objects
}
\examples{
# 1-parameter model
pkmod(pars_pk = c(CL = 10, V1 = 10))
}
