% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.poppkmod}
\alias{print.poppkmod}
\title{Print poppkmod}
\usage{
\method{print}{poppkmod}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object with class "pkmod" created by `poppkmod()`}

\item{...}{Additional arguments. Not used}

\item{digits}{Number of significant digits to print}
}
\value{
Prints description of pkmod
}
\description{
Print method for poppkmod objects
}
\examples{
data <- data.frame(ID = 1:5,
AGE = seq(20,60,by=10),
TBW = seq(60,80,by=5),
HGT = seq(150,190,by=10),
MALE = c(TRUE,TRUE,FALSE,FALSE,FALSE))
poppkmod(data, drug = "ppf", model = "eleveld")
poppkmod(data, drug = "ppf", model = "eleveld")
}
