% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{seqby}
\alias{seqby}
\title{Modified seq function to include bounds}
\usage{
seqby(from, to, by)
}
\arguments{
\item{from}{sequence starting value}

\item{to}{sequence end value}

\item{by}{increment of the sequence}
}
\description{
Modified seq function to include bounds
}
