% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sig_inf}
\alias{sig_inf}
\title{Return an infusion schedule defined by an Emax sigmoid curve.}
\usage{
sig_inf(beta, theta, ini, t0, E0, gamma, tmx = 10, bist = 50, delta = 1/6, ...)
}
\arguments{
\item{beta}{Parameters (c50, gamma) for Emax target function}

\item{theta}{Vector of PK-PD parameters (theta_PK, theta_PD)}

\item{ini}{Initial concentration values}

\item{t0}{Starting time}

\item{E0}{Effect at ce = 0. Used as starting point for sigmoid curve.}

\item{gamma}{Slope at c50}

\item{tmx}{Final time to evaluate infusion schedule to.}

\item{bist}{Target BIS, used to set the maximum effect value of the target sigmoid curve.}

\item{delta}{Time interval between TCI updates. Defaults to 1/6 minutes = 10 seconds.}
}
\description{
Return an infusion schedule defined by an Emax sigmoid curve.
}
