% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sigmoid_targetfn}
\alias{sigmoid_targetfn}
\title{Sigmoid target function starting at 93 and passing}
\usage{
sigmoid_targetfn(lpars, tms, bis0 = 93, ...)
}
\description{
Sigmoid target function starting at 93 and passing
}
