% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{tail_vec}
\alias{tail_vec}
\title{Extract last element or column}
\usage{
tail_vec(x)
}
\arguments{
\item{x}{Vector or matrix}
}
\description{
Function to extract the last element from a vector or the last column from a matrix
}
