% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{tci_documentation}
\alias{tci_documentation}
\title{tci_documentation}
\description{
Functions to implement target-controlled infusion algorithms
}
\details{
tci package documentation

TCI algorithms for 1- to 4- compartment PK models with IV administration.
}
