% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_algorithms.R
\name{tci_effect}
\alias{tci_effect}
\title{TCI algorithm for effect-site targeting}
\usage{
tci_effect(
  Cet,
  pkmod,
  dt = 1/6,
  ecmpt = NULL,
  tmax_search = 20,
  maxrt = 1200,
  grid_len = 1200,
  ...
)
}
\arguments{
\item{Cet}{Numeric vector of target effect-site concentrations.}

\item{pkmod}{PK model}

\item{dt}{Frequency of TCI updates. Default is 10 seconds expressed in terms of minutes = 1/6.}

\item{ecmpt}{Effect site compartment number}

\item{max_kR}{Maximum infusion rate of TCI pump. Defaults to 1200.}

\item{plasma_tol}{Maximum percent difference between predicted plasma concentration and target concentration permitted
in order to switch to plasma-targeting mode. Plasma-targeting mode is used primarily to increase computational efficiency
when the effect-site concentration is sufficiently stable and close to the target concentration.}

\item{effect_tol}{Maximum percent difference between predicted effect-site concentration and target concentration
permitted in order to switch to plasma-targeting mode.}
}
\description{
Function for calculating a TCI infusion schedule corresponding to a set of target concentrations.
This function makes use of formulas described by Shafer and Gregg (1992) in "Algorithms to rapidly achieve
and maintain stable drug concentrations at the site of drug effect with a computer-controlled infusion pump"
}
