% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.pkmod}
\alias{update.pkmod}
\title{Update method for pkmod}
\usage{
\method{update}{pkmod}(object, ...)
}
\arguments{
\item{object}{Object with class pkmod}

\item{...}{Updated values passed to object.}
}
\value{
Returns the pkmod object with elements replaced
}
\description{
Update parameters or initial values of a pkmod object
}
\examples{
# initial pkmod object
(my_mod <- pkmod(pars_pk = c(CL = 10, V1 = 10)))
# update a subset of parameters and initial values
update(my_mod, pars_pk = c(CL = 20, V1 = 2), init = 3, sigma_add = 1, log_response = TRUE)
}
