% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{validate_pkmod}
\alias{validate_pkmod}
\title{pkmod validation checks}
\usage{
validate_pkmod(x)
}
\arguments{
\item{x}{Object of class "pkmod"}
}
\value{
Returns a list with class "pkmod" if validation checks are passed. Returns an error if not.
}
\description{
Function to provide validation checks for a pkmod object
}
\examples{
validate_pkmod(init_pkmod(pars_pk = c(CL = 10, V1 = 10,Q2 = 4,V2=20)))
}
