% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{validate_poppkmod}
\alias{validate_poppkmod}
\title{Perform validation checks on a `poppkmod` object}
\usage{
validate_poppkmod(x)
}
\arguments{
\item{x}{Object with class "poppkmod" created by init_poppkmod}
}
\value{
`poppkmod` object or error
}
\description{
Perform validation checks on a `poppkmod` object created by `init_poppkmod`.
}
\examples{
data <- data.frame(ID = 1:5, AGE = seq(20,60,by=10), TBW = seq(60,80,by=5),
HGT = seq(150,190,by=10), MALE = c(TRUE,TRUE,FALSE,FALSE,FALSE))
validate_poppkmod(init_poppkmod(data, drug = "ppf", model = "eleveld"))
}
