% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{apply_poppk}
\alias{apply_poppk}
\title{Function to apply saved population PK or PK-PD models to a dataframe of patient values.}
\usage{
apply_poppk(patient_df, mod = c("marsh", "schnider", "eleveld"), ...)
}
\description{
Function to apply saved population PK or PK-PD models to a dataframe of patient values.
}
