% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{apply_targetfn}
\alias{apply_targetfn}
\title{Function to apply any specified target function to a PK-PD model and TCI algorithm}
\usage{
apply_targetfn(
  lp,
  tm,
  targetfn,
  prior_pk,
  prior_pd,
  pkmod = pkmod3cptm,
  pdmod = emax_eleveld,
  pdinv = inv_emax_eleveld,
  ...
)
}
\description{
Function to apply any specified target function to a PK-PD model and TCI algorithm
}
