% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{apply_targetfn}
\alias{apply_targetfn}
\title{Apply target function to a PK-PD model}
\usage{
apply_targetfn(
  lp,
  tm,
  targetfn,
  prior_pk,
  prior_pd,
  pkmod = pkmod3cptm,
  pdmod = emax_eleveld,
  pdinv = inv_emax_eleveld,
  ...
)
}
\arguments{
\item{lp}{Logged parameter values}

\item{tm}{Time values to evaluate}

\item{targetfn}{Target function}

\item{prior_pk}{Prior PK point estimates}

\item{prior_pd}{Prior PD point estimates}

\item{pkmod}{PK model to evaluate}

\item{pdmod}{PD model to evaluate}

\item{pdinv}{Inverse PD model}
}
\description{
Function to apply any specified target function to a PK-PD model
and TCI algorithm.
}
