% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{bayes_control}
\alias{bayes_control}
\title{Function to provide Bayesian closed-loop control to}
\usage{
bayes_control(
  targets,
  updates,
  prior,
  true_pars,
  pkmod = pkmod3cptm,
  pdmod = emax_eleveld,
  pdinv = inv_emax_eleveld,
  init0 = NULL,
  init_p = NULL,
  obs_tms = NULL,
  dt = 1/60,
  sim_starttm = 0,
  tci_alg = "effect"
)
}
\arguments{
\item{targets}{Dataframe with columns ("time","target")}

\item{updates}{Dataframe of times at which closed-loop updates should be conducted and
optional variable with logical values named 'full_data' indicating if full updates should
be used. Defaults to partial.}

\item{prior}{List with elements "mu" and "sig" specifying the prior mean and covariance
matrices for the logged parameter values.}

\item{obs_tms}{Times at which observations are collected. If null, observations will be
made at fixed intervals specified by 'dt'.}

\item{dt}{Interval between measurements.}
}
\description{
Function to provide Bayesian closed-loop control to
}
