% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eleveld_pk}
\alias{eleveld_pk}
\title{Eleveld et al. pharmacokinetic data}
\format{
A data frame with 1033 rows and 16 variables:
\describe{
  \item{ID}{Patient ID}
  \item{V1}{EB estimate of first compartment volume}
  \item{V2}{EB estimate of second compartment volume}
  \item{V3}{EB estimate of third compartment volume}
  \item{CL}{EB estimate of clearance for the first compartment}
  \item{Q2}{EB estimate of inter-compartmental clearance for second compartment}
  \item{Q3}{EB estimate of inter-compartmental clearance for third compartment}
  \item{AGE}{Patient's age (years)}
  \item{WGT}{Patient's weight (kg)}
  \item{HGT}{Patient's height (cm)}
  \item{M1F2}{Patient's sex: male = 1, female = 2}
  \item{PMA}{Patient's post-menstrual age. Assumed to be age + 40 weeks if not provided}
  \item{TECH}{Presence of concomitant anaesthetic techniques (Local anesthetic = 1, Opioids = 2)}
  \item{BMI}{Patient's BMI}
  \item{FFM}{Patient's fat-free mass (FFM)}
  \item{A1V2}{Sampling site: arterial sampling = 1, venous sampling = 2}
}
}
\usage{
data(eleveld_pk)
}
\description{
Empirical Bayes (EB) estimates of PK parameters made by the Eleveld et al (2018) PK-PD model.
}
\references{
Eleveld et al. (2018) British Journal of Anesthesia Vol. 120, 5:942-959
(\href{https://bjanaesthesia.org/article/S0007-0912(18)30051-5/abstract}{BJA})
}
\keyword{datasets}
