% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_mods.R
\name{emax_eleveld}
\alias{emax_eleveld}
\title{pars_emax <- c(c50 = 1.5, gamma = 4, e0 = 100, emx = 100)
ce_seq <- seq(0,4,0.1)
all.equal(inv_emax(emax(ce_seq, pars_emax), pars_emax), ce_seq)
Emax function for Eleveld (2018) model. The parameter gamma depends on whether ce <= c50.}
\usage{
emax_eleveld(ce, pars)
}
\arguments{
\item{ce}{Vector of effect-site concentrations.}

\item{pars}{Vector of parameter values in order (c50,gamma,gamma2,e0,emx).}
}
\description{
pars_emax <- c(c50 = 1.5, gamma = 4, e0 = 100, emx = 100)
ce_seq <- seq(0,4,0.1)
all.equal(inv_emax(emax(ce_seq, pars_emax), pars_emax), ce_seq)
Emax function for Eleveld (2018) model. The parameter gamma depends on whether ce <= c50.
}
