% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Function to evaluate the log likelihood given a set of logged parameter values and a set of observed BIS values.}
\usage{
log_likelihood(lpr, dat, fixed_lpr = NULL, pd_ix = 10, err_ix = 11)
}
\arguments{
\item{lpr}{logged PK-PD-error parameter values}

\item{dat}{data frame with columns c("time","bis") corresponding to observed time and bis values}

\item{fixed_lpr}{values used by PD function that are not updated.
These are concatenated to lpr[pd_ix] and the ordering may be important for use of the PD function.}

\item{pd_ix}{indices for PD function values to be updated.}

\item{err_ix}{index for standard deviation of residual error term}
}
\description{
Function to evaluate the log likelihood given a set of logged parameter values and a set of observed BIS values.
}
