% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{log_posterior_neg}
\alias{log_posterior_neg}
\title{Function to evaluate the negative log posterior given a set of logged parameter values and observed BIS values.}
\usage{
log_posterior_neg(lpr, dat, mu, sig, ...)
}
\arguments{
\item{lpr}{logged PK-PD-error parameter values}

\item{dat}{data frame with columns corresponding to  observed time and bis values}

\item{ivt}{infusion schedule}

\item{lhyper}{hyperparameter values to be passed to log_prior()}

\item{gamma}{gamma parameter of PD model (fixed in Eleveld model)}

\item{E0}{E0 parameter of PD model (fixed in Eleveld model)}
}
\description{
Function to evaluate the negative log posterior given a set of logged parameter values and observed BIS values.
}
