% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{log_posterior_neg}
\alias{log_posterior_neg}
\title{Function to evaluate the negative log posterior given a set of logged parameter values and observed BIS values.}
\usage{
log_posterior_neg(lpr, dat, mu, sig, ...)
}
\arguments{
\item{lpr}{logged PK-PD-error parameter values}

\item{dat}{data frame with columns corresponding to observed time and PD response values.}

\item{mu}{Mean of prior distribution.}

\item{sig}{Variance-covariance matrix of prior distribution.}

\item{...}{Arguments passed on to log-likelihood.}
}
\description{
Function to evaluate the negative log posterior given a set of logged parameter values and observed BIS values.
}
