% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{log_prior}
\alias{log_prior}
\title{Calculate logged prior value}
\usage{
log_prior(lpr, mu, sig)
}
\arguments{
\item{lpr}{log parameter values to evaluate}

\item{mu}{mean for model parameters and mean residual error}

\item{sig}{variance covariance matrix for model parameters}
}
\description{
Function to return the prior probability for a set of parameters
assuming a log-normal distribution.
}
