% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppk_mods.R
\name{marsh_poppk}
\alias{marsh_poppk}
\title{------------------------------------------------------------------------------------
Population PK and PK-PD functions --------------------------------------------------
------------------------------------------------------------------------------------
Marsh population PK model. Takes in a vector of patient weights and returns
a data frame of patient PK-PD parameters.
KE0 parameter set to 1.2 in accordance with Base Primea system and recommendations
from Absalom et al. (2009) "Pharmacokinetic models for propofol- Defining and
illuminating the devil in the detail"}
\usage{
marsh_poppk(df, rate = T)
}
\arguments{
\item{TBM}{Vector of patient total body mass in kg}
}
\description{
------------------------------------------------------------------------------------
Population PK and PK-PD functions --------------------------------------------------
------------------------------------------------------------------------------------
Marsh population PK model. Takes in a vector of patient weights and returns
a data frame of patient PK-PD parameters.
KE0 parameter set to 1.2 in accordance with Base Primea system and recommendations
from Absalom et al. (2009) "Pharmacokinetic models for propofol- Defining and
illuminating the devil in the detail"
}
