% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_mods.R
\name{pk_solution_3cpt_metab}
\alias{pk_solution_3cpt_metab}
\title{This function extends the function pk_basic_solution_3cpt_metab to a specified infusion schedule, rather than a single
infusion.}
\usage{
pk_solution_3cpt_metab(pars, ivt, init)
}
\arguments{
\item{pars}{Named vector of parameters for a 3-compartment model with effect-site.}

\item{ivt}{Infusion schedule given in the form of a named list
(e.g. list(list(begin = 0, end = 2, k_R = 1), list(begin = 4, end = 6, k_R = 1)))}

\item{init}{inital concentrations for the 4 compartments.}
}
\description{
This function extends the function pk_basic_solution_3cpt_metab to a specified infusion schedule, rather than a single
infusion.
}
