% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pkmod}
\alias{plot.pkmod}
\title{Plot method for 'pkmod'}
\usage{
\method{plot}{pkmod}(
  x,
  ...,
  inf,
  endtm = NULL,
  title = NULL,
  xlab = "Time",
  ylab = "Concentration",
  plot_pk = TRUE,
  ylab_resp = "Response"
)
}
\arguments{
\item{x}{An object with class pkmod.}

\item{...}{Arguments passed on to predict.pkmod. Could include changing initial concentrations or parameter values through pkmod_update. See ?predict.pkmod}

\item{inf}{An infusion schedule object with columns "begin","end","infrt".}

\item{endtm}{Final time to evaluate predictions}

\item{title}{Plot title}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{plot_pk}{Logical. Should PK concentrations be displayed in addition to PD, when PD is present.}

\item{ylab_resp}{y-axis label for PD component, if present.}
}
\value{
ggplot object displaying predicted concentrations for a pkmod object.
}
\description{
Will show predicted concentrations in compartments associated with an infusion schedule.
}
\examples{
# dosing schedule
dose <- create_inf(times = c(0,0.5,4,4.5,10), c(100,0,80,0,0))
# pkmod object
my_mod <- pkmod(pars_pk = c(CL = 15, V1 = 10, Q2 = 10, V2 = 20))
# plot predicted concentrations
plot(my_mod, inf = dose, ylab = "Concentration (mg/L)", xlab = "Minutes")
# plot with PD component
my_mod_pd <- pkmod(pars_pk = c(v1 = 8.995, v2 = 17.297, v3 = 120.963, cl = 1.382,
q2 = 0.919, q3 = 0.609, ke0 = 1.289),
pars_pd = c(c50 = 2.8, gamma = 1.47, gamma2 = 1.89, e0 = 93, emx = 93),
pdfn = emax, pdinv = emax_inv, ecmpt = 4)
plot(my_mod_pd, inf = dose, ylab = "Concentration (mg/L)", ylab_resp = "Bispectral Index",
xlab = "Minutes")
# plot with alternate PK and PD parameters
plot(my_mod_pd, inf = dose, pars = c(cl = 10), pars_pd = c(c50 = 2, emx = 100),
ylab = "Concentration (mg/L)", ylab_resp = "Bispectral Index", xlab = "Minutes")
}
