% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pkmod}
\alias{plot.pkmod}
\alias{plot.pdmod}
\alias{plot.tciinf}
\alias{plot.datasim}
\title{Plot pkmod object. Will show predicted concentrations in
compartments associated with an infusion schedule.}
\usage{
\method{plot}{pkmod}(x, ..., inf, npts = 1000, title = NULL)

\method{plot}{pdmod}(
  x,
  ...,
  pkmod,
  inf,
  pars_pd,
  pars_pk,
  npts = 1000,
  plot_pk = TRUE,
  title = NULL,
  ecmpt = NULL
)

\method{plot}{tciinf}(x, ..., title = NULL)

\method{plot}{datasim}(
  x,
  lpars_prior = NULL,
  lpars_update = NULL,
  lpars_fixed = NULL,
  pd_ix = 10,
  dt = 1/60,
  ...
)
}
\arguments{
\item{x}{An object with class datasim, created by function `gen_data`.}

\item{...}{\dots}

\item{inf}{An infusion schedule object with columns "begin","end","infrt".}

\item{npts}{Number of points used to evaluate predicted concentrations.}

\item{title}{Title of plot.}

\item{pars_pd}{Parameters used by pdmod.}

\item{pars_pk}{Parameters used by pkmod.}

\item{plot_pk}{Logical. Should PK concentrations be plotted alongside
the PD response. Defaults to TRUE.}

\item{ecmpt}{Effect-site compartment number. Defaults to the last
compartment concentration returned by pkmod.}

\item{lpars_prior}{Logged prior parameters for plotting predicted concentrations
alongside observed data.}

\item{lpars_update}{Logged posterior parameters for plotting predicted concentrations
alongside observed data.}

\item{lpars_fixed}{Logged parameters that are used by PK-PD model but not updated.}

\item{pd_ix}{Parameter vector of indices for PD parameters in arguments lpars_prior,
lpars_update.}

\item{dt}{Time interval used to predict PK-PD values.}
}
\description{
Plot pkmod object. Will show predicted concentrations in
compartments associated with an infusion schedule.

Plot method for PD models. User can provide a series of effect-site
concentrations and a PD model OR an infusion schedule with a PK-PD model

Plotting method for tciinf objects. Will work for outputs from "iterate_tci_grid" or "tci_pd".
Note: the tci targets function is left-continous. Consequently, it plots
final target value plotted will be a repetition of the preceeding value

Plot method for PD models. User can provide a series of effect-site
concentrations and a PD model OR an infusion schedule with a PK-PD model
}
