% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkpd_assist.R
\name{poppk_cov}
\alias{poppk_cov}
\title{Function to extract covariance from population pk or pk-pd models
This gives the covariance for the random effects about the fixed effect estimates.
The variance estimates for the residual error terms give the variance of error terms
within the population about the logged}
\usage{
poppk_cov(poppk = c("Schnider", "Eleveld"), pd = TRUE)
}
\description{
Function to extract covariance from population pk or pk-pd models
This gives the covariance for the random effects about the fixed effect estimates.
The variance estimates for the residual error terms give the variance of error terms
within the population about the logged
}
