% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppk_mods.R
\name{schnider_poppk}
\alias{schnider_poppk}
\title{Schnider population PK model}
\usage{
schnider_poppk(df, rate = FALSE, rand = FALSE)
}
\arguments{
\item{df}{data frame with variable names "AGE","TBM","HGT","MALE"}

\item{rate}{Logical. Should rate parameters be returned rather than clearance.
Defaults to FALSE}

\item{rand}{Logical. Should a vector of Monte Carlo samples be returned instead
of point estimates at patient covariate values. Defaults to FALSE.}
}
\description{
Evaluate Schnider population PK model at patient covariate values.
}
\examples{
marsh_poppk(data.frame(TBM = c(50,70,90)))
}
