% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sigmoid_targetfn}
\alias{sigmoid_targetfn}
\title{Sigmoid target function}
\usage{
sigmoid_targetfn(lpars, tms, bis0 = 93, ...)
}
\arguments{
\item{lpars}{Logged parameter values}

\item{tms}{Times to evaluate sigmoid function}

\item{bis0}{BIS value with no drug administered}

\item{...}{Arguments passed on to 'restrict_sigmoid' function}
}
\description{
Sigmoid target function
}
