% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_algorithms.R
\name{tci_comb}
\alias{tci_comb}
\title{Modified effect-site TCI algorithm that follows Jacobs (1993) suggestion of
switching to plasma-targeting when the plasma concentration is within 10\%
of the target and the effect-site concentration is within 0.5\% of the target.
The modification decreases computation time and prevents oscilatory behavior
in the effect-site concentrations.}
\usage{
tci_comb(Ct, pkmod, cptol = 0.1, cetol = 0.05, cp_cmpt = 1, ce_cmpt = 4, ...)
}
\description{
Modified effect-site TCI algorithm that follows Jacobs (1993) suggestion of
switching to plasma-targeting when the plasma concentration is within 10\%
of the target and the effect-site concentration is within 0.5\% of the target.
The modification decreases computation time and prevents oscilatory behavior
in the effect-site concentrations.
}
