% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_algorithms.R
\name{tci_comb}
\alias{tci_comb}
\title{Effect-site TCI algorithm with plasma targeting within small range of target}
\usage{
tci_comb(Ct, pkmod, cptol = 0.1, cetol = 0.05, cp_cmpt = 1, ce_cmpt = 4, ...)
}
\arguments{
\item{Ct}{Numeric vector of target effect-site concentrations.}

\item{pkmod}{PK model}

\item{cptol}{Percentage of effect-site concentration required to be within to switch
to plasma targeting.}
}
\description{
Modified effect-site TCI algorithm that follows Jacobs (1993) suggestion of
switching to plasma-targeting when the plasma concentration is within 10\%
of the target and the effect-site concentration is within 0.5\% of the target.
The modification decreases computation time and prevents oscillatory behavior
in the effect-site concentrations.
}
