% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{tci_documentation}
\alias{tci_documentation}
\title{tci_documentation}
\description{
Functions to implement target-controlled infusion algorithms
}
\details{
tci package documentation

This package contains functions to implement target-controlled
infusion (TCI) algorithms for compartmental PK models under intravenous administration.
TCI algorithms for plasma or effect-site targeting are included and can be
extended to pharmacodynamic responses. Custom PK-PD models and custom TCI
algorithms can be specified. Functions are provided to simulate responses from
PK/PK-PD models under open- or closed-loop control.
}
