% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tci_algorithms.R
\name{tci_pd}
\alias{tci_pd}
\title{Function to extend TCI grid to a set of PD targets}
\usage{
tci_pd(pdresp, tms, pkmod, pdmod, pars_pk, pars_pd, pdinv, ecmpt = NULL, ...)
}
\arguments{
\item{...}{Arguments to be passed on to 'tci'. These can include alternate TCI algorithms if desired.}
}
\description{
Function to extend TCI grid to a set of PD targets
}
